feat(tasks): Extract step executor and step notifier into separate files
