+++ src/screenplay-protractor.ts
@@ -1,4 +1,3 @@
 export * from './screenplay';
 export { Cast } from './serenity/stage';
-export { step } from './serenity/recording/step_annotation';
 export * from './serenity-protractor';
