+++ src/serenity/recording/index.ts
@@ -1,0 +1,4 @@
+export * from './step_annotation';
+export * from './step_description';
+export * from './step_executor';
+export * from './step_notifier';
